delete extra textures

+++ Bonus/Gameplay/music_bonus.c
@@ -6,7 +6,7 @@
 /*   By: adiri <adiri@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/12 06:52:22 by adiri             #+#    #+#             */
-/*   Updated: 2025/09/12 06:57:11 by adiri            ###   ########.fr       */
+/*   Updated: 2025/09/15 17:18:54 by adiri            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	stop_background_music(void)
 	t_gameplay	*gameplay;
 
 	gameplay = get_gameplay();
-	if (gameplay->start_game_sound)
+	if (gameplay->start_game_sound && gameplay->pid_sound > 0)
 		kill(gameplay->pid_sound, SIGKILL);
 	gameplay->start_game_sound = 0;
 }
@@ -43,17 +43,19 @@ void	*start_music_thread(void *arg)
 
 void	start_music(void)
 {
+	t_game		*game;
 	t_gameplay	*gameplay;
-	pthread_t	music_thread;
+	pthread_t	*music_thread;
 
-	gameplay = get_gameplay();
+	game = get_game();
+	gameplay = &game->gameplay;
+	music_thread = &game->music_thread;
 	if (gameplay->start_game_sound)
 		return ;
 	gameplay->start_game_sound = 1;
-	if (pthread_create(&music_thread, NULL, start_music_thread, NULL) != 0)
+	if (pthread_create(music_thread, NULL, start_music_thread, NULL) != 0)
 	{
 		report_error("pthread", "failed to create music thread");
-		leave_game(1);
+		gameplay->error_music = 1;
 	}
-	pthread_detach(music_thread);
 }
